Fix the returning type of `get_max_exponent`

+++ aonfp/detail/detail.hpp
@@ -51,7 +51,7 @@ constexpr T get_zero_mantissa_bitstring() {return static_cast<T>(0);};
 
 
 // range getter
-constexpr unsigned long get_max_exponent(const unsigned expopent_length) {return get_default_exponent_bias(expopent_length);}
+constexpr long get_max_exponent(const unsigned expopent_length) {return static_cast<long>(get_default_exponent_bias(expopent_length));}
 
 constexpr long get_min_exponent(const unsigned exponent_length) {return 1 - static_cast<long>(get_default_exponent_bias(exponent_length));}
